2002-05-06 Martin Weber <martweb@gmx.net>
	* filename.c (make_suffix): length_through_dot
	is typed to unsigned long for 64bit arch.
	(remove_suffix): Likewise.

	* curve.c (log_curve): Cast curve to unsigned long for 64bit
	arch.
	(log_entire_curve): Likewise.

	* autotrace.vcproj: Added output-dr2d.[ch].

	Message-ID: <3CD5A1C6.8717C29F@gmx.net>.
	Message-ID: <3CD5A2DB.2F467808@gmx.net>.

+++ curve.c
@@ -116,7 +116,7 @@ log_curve (curve_type curve, at_bool print_t)
 
   if (!log_file) return;
 
-  LOG1 ("curve id = %x:\n", (unsigned) curve);
+  LOG1 ("curve id = %x:\n", (unsigned long) curve);
   LOG1 ("  length = %u.\n", CURVE_LENGTH (curve));
   if (CURVE_CYCLIC (curve))
     LOG ("  cyclic.\n");
@@ -177,7 +177,7 @@ log_entire_curve (curve_type curve)
 
   if (!log_file) return;
 
-  LOG1 ("curve id = %x:\n", (unsigned) curve);
+  LOG1 ("curve id = %x:\n", (unsigned long) curve);
   LOG1 ("  length = %u.\n", CURVE_LENGTH (curve));
   if (CURVE_CYCLIC (curve))
     LOG ("  cyclic.\n");
